moved entire interface to a single place
