+++ services/director-v2/src/simcore_service_director_v2/api/routes/dynamic_scheduler.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, FastAPI, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from models_library.projects_nodes import NodeID
 from pydantic import BaseModel
 from servicelib.fastapi.long_running_tasks.client import (
@@ -14,22 +14,9 @@ from servicelib.fastapi.long_running_tasks.server import (
     start_task,
 )
 
-from ...core.settings import DynamicSidecarSettings
-from ...modules.dynamic_sidecar.api_client import DynamicSidecarClient
-from ...modules.dynamic_sidecar.scheduler import (
-    DynamicSidecarsScheduler,
-    push_outputs,
-    remove_containers,
-    remove_sidecar_proxy_docker_networks_and_volumes,
-    save_state,
-)
+from ...modules.dynamic_sidecar.scheduler import DynamicSidecarsScheduler
 from ...utils.routes import NoContentResponse
-from ..dependencies import get_app
-from ..dependencies.dynamic_sidecar import (
-    get_dynamic_sidecar_client,
-    get_dynamic_sidecar_scheduler,
-    get_dynamic_sidecar_settings,
-)
+from ..dependencies.dynamic_sidecar import get_dynamic_sidecar_scheduler
 
 
 class ObservationItem(BaseModel):
@@ -76,22 +63,20 @@ async def update_service_observation(
 async def delete_service_containers(
     node_uuid: NodeID,
     tasks_manager: TasksManager = Depends(get_tasks_manager),
-    dynamic_sidecar_client: DynamicSidecarClient = Depends(get_dynamic_sidecar_client),
-    app: FastAPI = Depends(get_app),
+    dynamic_sidecars_scheduler: DynamicSidecarsScheduler = Depends(
+        get_dynamic_sidecar_scheduler
+    ),
 ):
     async def _task_remove_service_containers(
-        task_progress: TaskProgress,
-        app: FastAPI,
-        node_uuid: NodeID,
-        dynamic_sidecar_client: DynamicSidecarClient,
+        task_progress: TaskProgress, node_uuid: NodeID
     ) -> None:
         async def _progress_callback(
             message: ProgressMessage, percent: ProgressPercent, _: TaskId
         ) -> None:
             task_progress.update(message=message, percent=percent)
 
-        await remove_containers(
-            app, node_uuid, dynamic_sidecar_client, _progress_callback
+        await dynamic_sidecars_scheduler.remove_service_containers(
+            node_uuid=node_uuid, progress_callback=_progress_callback
         )
 
     try:
@@ -99,9 +84,7 @@ async def delete_service_containers(
             tasks_manager,
             task=_task_remove_service_containers,
             unique=True,
-            app=app,
             node_uuid=node_uuid,
-            dynamic_sidecar_client=dynamic_sidecar_client,
         )
         return task_id
     except TaskAlreadyRunningError as e:
@@ -122,30 +105,29 @@ async def delete_service_containers(
 async def save_service_state(
     node_uuid: NodeID,
     tasks_manager: TasksManager = Depends(get_tasks_manager),
-    dynamic_sidecar_client: DynamicSidecarClient = Depends(get_dynamic_sidecar_client),
-    app: FastAPI = Depends(get_app),
+    dynamic_sidecars_scheduler: DynamicSidecarsScheduler = Depends(
+        get_dynamic_sidecar_scheduler
+    ),
 ):
     async def _task_save_service_state(
         task_progress: TaskProgress,
-        app: FastAPI,
         node_uuid: NodeID,
-        dynamic_sidecar_client: DynamicSidecarClient,
     ) -> None:
         async def _progress_callback(
             message: ProgressMessage, percent: ProgressPercent, _: TaskId
         ) -> None:
             task_progress.update(message=message, percent=percent)
 
-        await save_state(app, node_uuid, dynamic_sidecar_client, _progress_callback)
+        await dynamic_sidecars_scheduler.save_service_state(
+            node_uuid=node_uuid, progress_callback=_progress_callback
+        )
 
     try:
         task_id = start_task(
             tasks_manager,
             task=_task_save_service_state,
             unique=True,
-            app=app,
             node_uuid=node_uuid,
-            dynamic_sidecar_client=dynamic_sidecar_client,
         )
         return task_id
     except TaskAlreadyRunningError as e:
@@ -166,30 +148,28 @@ async def save_service_state(
 async def push_service_outputs(
     node_uuid: NodeID,
     tasks_manager: TasksManager = Depends(get_tasks_manager),
-    dynamic_sidecar_client: DynamicSidecarClient = Depends(get_dynamic_sidecar_client),
-    app: FastAPI = Depends(get_app),
+    dynamic_sidecars_scheduler: DynamicSidecarsScheduler = Depends(
+        get_dynamic_sidecar_scheduler
+    ),
 ):
     async def _task_push_service_outputs(
-        task_progress: TaskProgress,
-        app: FastAPI,
-        node_uuid: NodeID,
-        dynamic_sidecar_client: DynamicSidecarClient,
+        task_progress: TaskProgress, node_uuid: NodeID
     ) -> None:
         async def _progress_callback(
             message: ProgressMessage, percent: ProgressPercent, _: TaskId
         ) -> None:
             task_progress.update(message=message, percent=percent)
 
-        await push_outputs(app, node_uuid, dynamic_sidecar_client, _progress_callback)
+        await dynamic_sidecars_scheduler.push_service_outputs(
+            node_uuid=node_uuid, progress_callback=_progress_callback
+        )
 
     try:
         task_id = start_task(
             tasks_manager,
             task=_task_push_service_outputs,
             unique=True,
-            app=app,
             node_uuid=node_uuid,
-            dynamic_sidecar_client=dynamic_sidecar_client,
         )
         return task_id
     except TaskAlreadyRunningError as e:
@@ -210,19 +190,15 @@ async def push_service_outputs(
 async def delete_service_docker_resources(
     node_uuid: NodeID,
     tasks_manager: TasksManager = Depends(get_tasks_manager),
-    app: FastAPI = Depends(get_app),
-    dynamic_sidecar_settings: DynamicSidecarSettings = Depends(
-        get_dynamic_sidecar_settings
+    dynamic_sidecars_scheduler: DynamicSidecarsScheduler = Depends(
+        get_dynamic_sidecar_scheduler
     ),
 ):
     async def _task_cleanup_service_docker_resources(
-        task_progress: TaskProgress,
-        app: FastAPI,
-        node_uuid: NodeID,
-        dynamic_sidecar_settings: DynamicSidecarSettings,
+        task_progress: TaskProgress, node_uuid: NodeID
     ) -> None:
-        await remove_sidecar_proxy_docker_networks_and_volumes(
-            task_progress, app, node_uuid, dynamic_sidecar_settings
+        await dynamic_sidecars_scheduler.remove_service_sidecar_proxy_docker_networks_and_volumes(
+            task_progress=task_progress, node_uuid=node_uuid
         )
 
     try:
@@ -230,9 +206,7 @@ async def delete_service_docker_resources(
             tasks_manager,
             task=_task_cleanup_service_docker_resources,
             unique=True,
-            app=app,
             node_uuid=node_uuid,
-            dynamic_sidecar_settings=dynamic_sidecar_settings,
         )
         return task_id
     except TaskAlreadyRunningError as e:
